Add basic interpolated string nullable handling and tests.
